refactor: 타입 정의 파일 구조 정리 및 불필요한 타입 제거
