+++ src/shared/types/response/index.ts
@@ -1,6 +1,6 @@
-export * from './cart.types';
-export * from './common.types';
-export * from './products.types';
-export * from './recipe.types';
-export * from './token.types';
-export * from './user.types';
+export * from './cart';
+export * from './common';
+export * from './products';
+export * from './recipe';
+export * from './token';
+export * from './user';
