+++ src/shared/types/response/products.ts
