+++ src/shared/types/response/token.ts
