+++ src/shared/types/response/user.ts
